+++ spec/fixtures/vcr_cassettes/spec/requests/search/cda/defendant_by_name_spec.yml
@@ -2,15 +2,17 @@
 http_interactions:
 - request:
     method: get
-    uri: https://laa-court-data-api-dev.apps.live.cloud-platform.service.justice.gov.uk/v2/defendants?nino=HX685369B
+    uri: http://localhost:9292/api/internal/v2/prosecution_cases?filter%5Bdate_of_birth%5D=1990-01-01&filter%5Bname%5D=trever%20glover
     body:
       encoding: US-ASCII
       string: ''
     headers:
-      Authorization:
-      - Basic <Basic>
       Accept:
       - application/json
+      Content-Type:
+      - application/json
+      Authorization:
+      - Bearer <BEARER_TOKEN>
       Accept-Encoding:
       - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
       User-Agent:
@@ -20,19 +22,35 @@ http_interactions:
       code: 200
       message: OK
     headers:
-      Date:
-      - Thu, 14 Apr 2022 10:10:14 GMT
+      X-Frame-Options:
+      - SAMEORIGIN
+      X-Xss-Protection:
+      - '0'
+      X-Content-Type-Options:
+      - nosniff
+      X-Permitted-Cross-Domain-Policies:
+      - none
+      Referrer-Policy:
+      - strict-origin-when-cross-origin
+      X-Request-Id:
+      - 9bf8d3dc-9bc0-4b22-a899-23f132444ab5
+      Cache-Control:
+      - no-store
+      Www-Authenticate:
+      - Bearer realm="Doorkeeper", error="invalid_token", error_description="The access
+        token is invalid"
       Content-Type:
-      - application/json
+      - text/html
+      X-Runtime:
+      - '0.028687'
+      Server-Timing:
+      - start_processing.action_controller;dur=0.01, sql.active_record;dur=0.46, halted_callback.action_controller;dur=0.01,
+        process_action.action_controller;dur=2.66
       Content-Length:
-      - '5183'
-      Connection:
-      - keep-alive
-      Strict-Transport-Security:
-      - max-age=15724800; includeSubDomains
+      - '0'
     body:
       encoding: UTF-8
-      string: '{"defendant_summaries":[{"id":"a23d0911-4806-42eb-b910-e2bdcf3018d9","national_insurance_number":"HX685369B","arrest_summons_number":"FPV3XYUJNPIB","name":"Wendie
+      string: '{"total_results": 1, "results": [{ "prosecution_case_reference": "TEST12345", "defendant_summaries":[{"id":"a23d0911-4806-42eb-b910-e2bdcf3018d9","national_insurance_number":"HX685369B","arrest_summons_number":"FPV3XYUJNPIB","name":"Wendie
         Bogisiche Lowe","first_name":"Wendie","middle_name":"Bogisiche","last_name":"Lowe","date_of_birth":"2000-12-27","proceedings_concluded":false,"representation_order":{"laa_application_reference":null,"effective_start_date":null,"effective_end_date":null,"laa_contract_number":null},"offence_summaries":[{"id":"6fc6b4a7-131c-464b-a9fe-7440e49311b9","code":"PT00011","order_index":1,"title":"Possessing
         anything with intent to destroy or damage property","legislation":"Sexual
         Offences Act 2003 s. 14","wording":"Random string","arrest_date":"2019-10-17","charge_date":"2019-10-17","mode_of_trial":"Indictable
@@ -51,40 +69,6 @@ http_interactions:
         criminal damage","legislation":"Sexual Offences Act 1956 s.19","wording":"Random
         string","arrest_date":"2019-10-17","charge_date":"2019-10-17","mode_of_trial":"Summary
         offence","start_date":"2019-10-17","proceedings_concluded":false,"plea":{"date":"2019-10-23","value":"NO_PLEA","originating_hearing_id":"89fc86c1-a92c-4250-b673-a4e6d5421477","offence_id":null,"application_id":null,"delegated_powers":null},"verdict":{"date":null,"type":null,"originating_hearing_id":null,"offence_id":null},"laa_application":{"reference":"Z10000199","id":"c78edb29-98b7-40d2-8bbe-ed02ba6792d0","status_code":"AP","description":"FAKE
-        NEWS","status_date":"2022-04-08","effective_start_date":null,"effective_end_date":null,"contract_number":null}}],"prosecution_case_reference":"TEST12345"}]}'
-  recorded_at: Thu, 14 Apr 2022 10:10:14 GMT
-- request:
-    method: get
-    uri: https://laa-court-data-api-dev.apps.live.cloud-platform.service.justice.gov.uk/v2/defendants?nino=GP999999B
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      Authorization:
-      - Basic <Basic>
-      Accept:
-      - application/json
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      User-Agent:
-      - Ruby
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Date:
-      - Thu, 14 Apr 2022 10:10:17 GMT
-      Content-Type:
-      - application/json
-      Content-Length:
-      - '26'
-      Connection:
-      - keep-alive
-      Strict-Transport-Security:
-      - max-age=15724800; includeSubDomains
-    body:
-      encoding: UTF-8
-      string: '{"defendant_summaries":[]}'
-  recorded_at: Thu, 14 Apr 2022 10:10:17 GMT
-recorded_with: VCR 6.1.0
+        NEWS","status_date":"2022-04-08","effective_start_date":null,"effective_end_date":null,"contract_number":null}}],"prosecution_case_reference":"TEST12345"}]}]}'
+  recorded_at: Fri, 09 May 2025 15:02:58 GMT
+recorded_with: VCR 6.3.1
