--- spec/fixtures/vcr_cassettes/spec/requests/search/defendant_by_name_spec.yml
@@ -1,92 +1,0 @@
----
-http_interactions:
-- request:
-    method: get
-    uri: https://uat.court-data-adaptor.service.justice.gov.uk/api/internal/v1/prosecution_cases?filter%5Bdate_of_birth%5D=1990-01-01&filter%5Bname%5D=trever%20glover&include=defendants
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      User-Agent:
-      - Faraday v1.10.0
-      Content-Type:
-      - application/vnd.api+json
-      Accept:
-      - application/vnd.api+json
-      Authorization:
-      - Bearer <BEARER_TOKEN>
-      Accept-Encoding:
-      - gzip,deflate
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Date:
-      - Tue, 23 Mar 2021 13:07:45 GMT
-      Content-Type:
-      - application/vnd.api+json; charset=utf-8
-      Transfer-Encoding:
-      - chunked
-      Connection:
-      - keep-alive
-      X-Frame-Options:
-      - SAMEORIGIN
-      X-Xss-Protection:
-      - 1; mode=block
-      X-Content-Type-Options:
-      - nosniff
-      X-Download-Options:
-      - noopen
-      X-Permitted-Cross-Domain-Policies:
-      - none
-      Referrer-Policy:
-      - strict-origin-when-cross-origin
-      Laa-Transaction-Id:
-      - '0943e894eb8768c1b99db0fda5fc7a21'
-      Vary:
-      - Accept,Accept-Encoding
-      Etag:
-      - W/"000ab64177e2f66d2fed5b42d019158e"
-      Cache-Control:
-      - max-age=0, private, must-revalidate
-      X-Request-Id:
-      - '0943e894eb8768c1b99db0fda5fc7a21'
-      X-Runtime:
-      - '5.194981'
-      Content-Encoding:
-      - gzip
-      Strict-Transport-Security:
-      - max-age=15724800; includeSubDomains
-    body:
-      encoding: UTF-8
-      base64_string: |
-        H4sIACHoWWAAA9SXbWsjNxDHv8qxrzNFGj3nXeCueZO2UEILPYIZSaOLwVmH
-        9boQQr57Z53e5TlN7IQSsLGX1Uizf/0089/LrtJI3f7Xy25eu/2uWFspNwXk
-        HIJNFoFabhCcsqxSDppKt9eNF+cso8+H5YrLepwv+1mhFa/kFo3jMM/rUS72
-        Lx+MmA3ceOC+TOFJHX5OGvHo4KC72usGXtA0cHU6P98EVxnbV+rH66s7ibak
-        nfM1QUm1glVGA0m6wDb5Gh2yfG8SvTXT1Yksdco0zPtvs9X67Ez+8SMLBBM4
-        YmIoLlqwhjJE1RzEkENWFTHodLPAwwmv9v6dCEU1U6No6KMByxggKZ+gBZKs
-        fcBi3LMT3Ur4rfLcNb1NVlc3c5hSkvEMrHwRajRCbhGhaXJILvms35galZRO
-        +tXUeMNsahBqvJGH9QIMJWMhI3t0TmvV/E7URM6+6EoQIspusEFIpogcRuXk
-        iq1e2202+7baGQNKtglUIQLbWNbAqkB7nUxMzNrVd1AbX622KzpUE4RFDAqk
-        sAQg5RhqKOhK0jpHtZPa2TtnFJPQVpwUASkHUg08KLROsE252rir2uSTV4UL
-        5FAD2BwrpIgGQmVTEqaGtb2D2ubVaqsYGqmMoKITLOKERS5yHIulVmzJ5MJu
-        apM25Iyc8pLl8GAJEGNgcIXRYFEqOdpVbQwl+xADBOubHNBaIGarpmqkTTBO
-        jm18B7Xtq9XOpRmHKgMye9E4JsiUpHZ6hyFgzfn2IdxG7dykiYnQ5DnLKVcG
-        srYNlInGcm2eat1W7ZO9bt6XxbpyfYOOel/zns6mEccD/83Dp8PFUn5kkKzP
-        s2Wb5fkwnsp9nZICpeUjN/uN8rSYzfvVeiDZn1m/PssSuN+vFwtZYhh4NW6e
-        T3box83u+OcjCT8jGu9srVda+Tg1voHPJZL7cbPCbDlUHmZTLhNtCiWFCDod
-        K7W/+fyklPpLwh6A8xp7dL6UVL9vSVmvxmW9mK0kg/XE6P7X7qA7eQy4ZWtT
-        /o/QwFz15lwntnL2nPSsRLmCr9KfY3JS+/XN5vyYZwPBZqtE0OXwjfr5arPk
-        zQrfD94vqM2v9/b3XpDMd/VQmQdT7eYhZY2bcrB1v36CyT/niyYE/D9QXsc8
-        w+P1gMfJe6HF+i/yXomdFUPB2XrxFoWlJiBPNQFBK25SoQtZto9h96N5ctYp
-        StY0eSIr1R2i9x5KS4VSSlRC3g7bSauXwLibNb0D49Z25gkYDxbzC/p0ePTb
-        H19+/0govtB/vjGKJPtoUWtAp6SqeGKIhHJpTYsJfQv1WRTFj0udJCn2CqWf
-        Fp+lYZN4xWyDD1YKSXbvi+Juvv0Oilt7vSdQ/CyPRYsPyOIL3fkbsxg1kfOK
-        gA1JWwri+8h4Dc44MakKQ6JnWZQok1QjiFUwtGKw5J1bXmPEQFbT0Lgcyvuy
-        uNtbzR0Wt3bCT7D4RTrz+AFRfOGryxuj2KJUExMdBLTi2p0UgaSshupcVU0X
-        Cs48i6KSGCtOMiJLfKlZjGWRzay2kECQPb13h97plW+D4snVPwAAAP//AwDQ
-        JB3dyBQAAA==
-  recorded_at: Tue, 23 Mar 2021 13:07:45 GMT
-recorded_with: VCR 6.0.0
