--- spec/fixtures/vcr_cassettes/spec/requests/search/defendant_by_reference_spec.yml
@@ -1,65 +1,0 @@
----
-http_interactions:
-- request:
-    method: get
-    uri: http://localhost:9292/api/internal/v1/prosecution_cases?filter%5Bnational_insurance_number%5D=YE744478B&include=defendants,defendants.offences
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      User-Agent:
-      - Faraday v1.10.0
-      Content-Type:
-      - application/vnd.api+json
-      Accept:
-      - application/vnd.api+json
-      Authorization:
-      - Bearer <BEARER_TOKEN>
-      Accept-Encoding:
-      - gzip,deflate
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      X-Frame-Options:
-      - SAMEORIGIN
-      X-Xss-Protection:
-      - 1; mode=block
-      X-Content-Type-Options:
-      - nosniff
-      X-Download-Options:
-      - noopen
-      X-Permitted-Cross-Domain-Policies:
-      - none
-      Referrer-Policy:
-      - strict-origin-when-cross-origin
-      Content-Type:
-      - application/vnd.api+json; charset=utf-8
-      Etag:
-      - W/"401d9efdf72a1258e31119f18334e6be"
-      Cache-Control:
-      - max-age=0, private, must-revalidate
-      X-Request-Id:
-      - 652e1f56-6073-4c9b-a8a5-3d575fec9bea
-      X-Runtime:
-      - '0.067451'
-      Transfer-Encoding:
-      - chunked
-    body:
-      encoding: UTF-8
-      string: '{"data":[{"id":"db3c94f1-bc26-4674-9385-c0df45e48493","type":"prosecution_cases","attributes":{"prosecution_case_reference":"MOGUERBXIZ"},"relationships":{"defendants":{"data":[{"id":"d95fb19b-7af4-46a5-bc73-79cfe9eb511c","type":"defendants"},{"id":"2a9e8891-bfae-465d-95d0-483b0970e269","type":"defendants"},{"id":"69813351-658f-4d75-b6b6-3e8811be0be1","type":"defendants"},{"id":"0100d486-b3a2-4988-81aa-0e1c99b98ee2","type":"defendants"}]}}}],"included":[{"id":"d95fb19b-7af4-46a5-bc73-79cfe9eb511c","type":"defendants","attributes":{"name":"Josefa
-        Franecki","date_of_birth":"1961-06-15","national_insurance_number":"YE744478B","arrest_summons_number":"US7FIK88TA9X","maat_reference":null},"relationships":{"offences":{"data":[{"id":"43b560f0-8aa8-4de5-b0e5-aab7b8078d1e","type":"offences"}]}}},{"id":"2a9e8891-bfae-465d-95d0-483b0970e269","type":"defendants","attributes":{"name":"Weldon
-        Leffler Deckow","date_of_birth":"1962-06-18","national_insurance_number":"GG367471A","arrest_summons_number":"NYC1FXJTLWFE","maat_reference":null},"relationships":{"offences":{"data":[{"id":"4e8f8627-60b1-4f14-a7ca-49d01914b160","type":"offences"}]}}},{"id":"69813351-658f-4d75-b6b6-3e8811be0be1","type":"defendants","attributes":{"name":"Amada
-        Cruickshank Abshire","date_of_birth":"1977-08-20","national_insurance_number":"HE708741C","arrest_summons_number":"UEDOC8A64W7I","maat_reference":null},"relationships":{"offences":{"data":[{"id":"c1c8fcdf-c7f2-46f4-8fe4-f86400eac211","type":"offences"}]}}},{"id":"0100d486-b3a2-4988-81aa-0e1c99b98ee2","type":"defendants","attributes":{"name":"Jammy  Dodger","date_of_birth":"1987-05-21","national_insurance_number":"JC123456A","arrest_summons_number":"WWVHYGO7CMZ1","maat_reference":null},"relationships":{"offences":{"data":[{"id":"9f14c8ce-78e7-41e5-8ceb-4e7872b86c51","type":"offences"}]}}},{"id":"43b560f0-8aa8-4de5-b0e5-aab7b8078d1e","type":"offences","attributes":{"code":"Random
-        string","order_index":1,"title":"Possession of offensive weapon","mode_of_trial":"Random
-        string"}},{"id":"4e8f8627-60b1-4f14-a7ca-49d01914b160","type":"offences","attributes":{"code":"Random
-        string","order_index":1,"title":"Indecent assault on a man","mode_of_trial":"Random
-        string"}},{"id":"c1c8fcdf-c7f2-46f4-8fe4-f86400eac211","type":"offences","attributes":{"code":"Random
-        string","order_index":1,"title":"Assaulting prison officer whilst possessing
-        firearm etc.","mode_of_trial":"Random string"}},{"id":"9f14c8ce-78e7-41e5-8ceb-4e7872b86c51","type":"offences","attributes":{"code":"Random
-        string","order_index":1,"title":"Acquisition by or supply of firearms to person
-        denied them","mode_of_trial":"Random string"}}]}'
-    http_version: null
-  recorded_at: Mon, 20 Apr 2020 13:35:19 GMT
-recorded_with: VCR 5.1.0
