+++ spec/requests/search/cda/defendant_by_name_spec.rb
@@ -41,7 +41,7 @@ RSpec.describe 'Defendant by name and dob search', type: :request do
 
       it 'assigns array of results' do
         expect(assigns(:results))
-          .to include(an_instance_of(CdApi::Defendant))
+          .to include(an_instance_of(Cda::Defendant))
       end
 
       it 'renders searches/_results' do
