+++ spec/requests/search/cda/defendant_by_reference_spec.rb
@@ -38,7 +38,7 @@ RSpec.describe 'Defendant by reference search', type: :request do
 
       it 'assigns array of results' do
         expect(assigns(:results))
-          .to include(an_instance_of(CdApi::Defendant))
+          .to include(an_instance_of(Cda::Defendant))
       end
 
       it 'renders searches/_results' do
