+++ spec/requests/search/cda/searches_spec.rb
